requirements.txt
pydantic>=2.8
pytest>=8.0
python-dotenv==1.0.0
requests>=2.31.0
psycopg2-binary
sqlalchemy==2.0.40
alembic==1.13.1
httpx>=0.28.0